feat: add warning when controlled state changes (#568)
Closes #566

<!--
Thanks for your interest in the project. Bugs filed and PRs submitted are appreciated!

Please make sure that you are familiar with and follow the Code of Conduct for
this project (found in the CODE_OF_CONDUCT.md file).

Also, please make sure you're familiar with and follow the instructions in the
contributing guidelines (found in the CONTRIBUTING.md file).

If you're new to contributing to open source projects, you might find this free
video course helpful: http://kcd.im/pull-request

Please fill out the information below to expedite the review and (hopefully)
merge of your pull request!
-->

<!-- What changes are being made? (What feature/bug is being fixed here?) -->

**What**:

Added a warning (mostly stolen from React itself) when a Downshift prop switches from controlled to uncontrolled (or vice versa) while the component is mounted.

**Why**:

See https://github.com/paypal/downshift/issues/566 for bug report - if you accidentally pass `selectedItem={undefined}` to a Downshift component, it will behave strangely (as it becomes uncontrolled) and it took me quite a while to figure out why.

**How**:

Checked all of the props that can be controlled in `componentDidUpdate` (only if `NODE_ENV !== "production"`). Log an error if one of them changes to or from `undefined`.

<!-- Have you done all of these things?  -->

**Checklist**:

<!-- add "N/A" to the end of each line that's irrelevant to your changes -->

<!-- to check an item, place an "x" in the box like so: "- [x] Documentation" -->

- [ ] Documentation N/A
- [x] Tests
- [x] Ready to be merged
      <!-- In your opinion, is this ready to be merged as soon as it's reviewed? -->
- [x] Added myself to contributors table
      <!-- this is optional, see the contributing guidelines for instructions -->

<!-- feel free to add additional comments -->
Unfortunately, I wasn't able to use all of the workflow scripts due to https://github.com/paypal/downshift/issues/567 - let me know if there's any scripts I need to run that don't rely on all of the tests passing.